# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lumen API",
        "description": "API for the Lumen project",
        "version": "1.0.0"
    },
    "paths": {
...
        "/criterios-evaluacion": {
            "get": {
                "tags": [
                    "Criterios de Evaluación"
                ],
                "summary": "Listar todos los criterios de evaluación docente",
                "operationId": "d63c389142959c27bb04aab3a01f794a",
                "responses": {
                    "200": {
                        "description": "Listado de criterios de evaluación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CriterioEvaluacion"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al obtener los criterios"
                    }
                }
            },
            "post": {
                "tags": [
                    "Criterios de Evaluación"
                ],
                "summary": "Crear un nuevo criterio de evaluación docente",
                "operationId": "a0e4593855bbdcbb124ddbed78a70743",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_rubro",
                                    "descripcion"
                                ],
                                "properties": {
                                    "id_rubro": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "El docente demuestra dominio del tema"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Criterio de evaluación creado exitosamente"
                    },
                    "400": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error al crear el criterio de evaluación"
                    }
                }
            }
        },
        "/criterios-evaluacion/{id}": {
            "get": {
                "tags": [
                    "Criterios de Evaluación"
                ],
                "summary": "Mostrar un criterio de evaluación docente específico",
                "operationId": "8631eb6dc887cce9499490ceed8f6474",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del criterio de evaluación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Criterio de evaluación encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CriterioEvaluacion"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Criterio de evaluación no encontrado"
                    },
                    "500": {
                        "description": "Error al obtener el criterio"
                    }
                }
            },
            "put": {
                "tags": [
                    "Criterios de Evaluación"
                ],
                "summary": "Actualizar un criterio de evaluación docente existente",
                "operationId": "9080d7c2c13ca6e2228566a1e8271577",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del criterio de evaluación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_rubro",
                                    "descripcion"
                                ],
                                "properties": {
                                    "id_rubro": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "El docente demuestra un excelente dominio del tema"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Criterio de evaluación actualizado exitosamente"
                    },
                    "400": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error al actualizar el criterio de evaluación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Criterios de Evaluación"
                ],
                "summary": "Eliminar un criterio de evaluación docente",
                "operationId": "a1534a567e24d1cccf4a4ebaf15793e5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del criterio de evaluación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Criterio de evaluación eliminado exitosamente"
                    },
                    "400": {
                        "description": "Error al eliminar el criterio de evaluación"
                    },
                    "500": {
                        "description": "Error al eliminar el criterio de evaluación"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CriterioEvaluacion": {
                "title": "CriterioEvaluacion",
                "description": "Modelo de Criterio de Evaluación",
                "properties": {
                    "id_evacriterio": {
                        "title": "ID del Criterio de Evaluación",
                        "description": "Identificador único del criterio de evaluación",
                        "type": "integer",
                        "readOnly": "true",
                        "example": 1
                    },
                    "id_rubro": {
                        "title": "ID del Rubro",
                        "description": "Identificador único del rubro asociado",
                        "type": "integer",
                        "example": 1
                    },
                    "descripcion": {
                        "title": "Descripción",
                        "description": "Descripción del criterio de evaluación",
                        "type": "string",
                        "maxLength": 255,
                        "example": "El docente demuestra dominio del tema"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "CriterioEvaluacion"
                }
            },
...
        }
    },
...
}